+++ src/headers/ImportDir.cpp
@@ -7,10 +7,10 @@
 using namespace Pepper;
 
 // each class related to regular imports must see the same `diskToMemDiff` value
-int ImportDir::s_diskToMemDiff = 0;
-template<> int *ImportDescriptor::s_pDiskToMemDiff = &ImportDir::s_diskToMemDiff;
-template<> int *ImportThunk32::s_pDiskToMemDiff    = &ImportDir::s_diskToMemDiff;
-template<> int *ImportThunk64::s_pDiskToMemDiff    = &ImportDir::s_diskToMemDiff;
+size_t ImportDir::s_diskToMemDiff = 0;
+template<> size_t *ImportDescriptor::s_pDiskToMemDiff = &ImportDir::s_diskToMemDiff;
+template<> size_t *ImportThunk32::s_pDiskToMemDiff    = &ImportDir::s_diskToMemDiff;
+template<> size_t *ImportThunk64::s_pDiskToMemDiff    = &ImportDir::s_diskToMemDiff;
 
 ImportDir::ImportDir(const PeFile &pe, const FileBytes &fbytes, const DataDirectoryEntry &dde)
 : IDirectory(pe, fbytes, dde)
@@ -20,9 +20,9 @@ ImportDir::ImportDir(const PeFile &pe, const FileBytes &fbytes, const DataDirect
         // append non-null IMPORT_DESCRIPTOR to m_descriptors vector
         char null[sizeof(IMAGE_IMPORT_DESCRIPTOR)] = {0};
         IMAGE_IMPORT_DESCRIPTOR *pid = (PIMAGE_IMPORT_DESCRIPTOR)dir();
-        int i=0;
+        size_t i=0;
         while (memcmp(&pid[i], &null, sizeof(null))) {
-            m_descriptors.emplace_back(pe, fbytes, dirOffset() + (i * (int)sizeof(null)));
+            m_descriptors.emplace_back(pe, fbytes, dirOffset() + (i * sizeof(null)));
             i++;
         }
     }
